backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.8.3
pydantic==2.7.4
pydantic-core==2.18.4
python-multipart==0.0.6
